주차장 정보 보기 구현 중 (메핑 오류 수정 중)

+++ src/main/java/com/el/parkingLot/dto/parking/ParkInfoDto.java
@@ -10,7 +10,7 @@ import lombok.*;
 @ToString
 @NoArgsConstructor
 public class ParkInfoDto {
-    private int pcInfo;
+    private Long pcInfo;
 
     // 외래키
     private MemberDto memberDto = new MemberDto();
